requirement.txt
numpy==1.23.5
nltk==3.8.1
tensorflow==2.12.0
Keras==2.12.0
python-dotenv==1.0.0
ibm-watson==7.0.0
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.3
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.1
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.4
requests==2.29.0
seaborn==0.12.2
six==1.16.0
spotipy==2.23.0
tzdata==2023.3
urllib3==1.26.15
flask
boto3